chore(jaeger): support Ipv6 address for sync uploader
